// src/types/Carrera.ts
export type Carrera =
  | 'INGENIERIA_QUIMICA'
  | 'INGENIERIA_MECANICA'
  | 'INGENIERIA_ELECTRICA'
  | 'INGENIERIA_ENERGIA'
  | 'INGENIERIA_MECATRONICA'
  | 'BIOINGENIERIA'
  | 'INGENIERA_INDUSTRIAL'
  | 'INGENIERIA_AMBIENTAL'
  | 'INGENIERIA_CIVIL'
  | 'SISTEMAS_DE_INFORMACION'
  | 'COMPUTER_SCIENCE'
  | 'DATA_SCIENCE'
  | 'CIBERSEGURIDAD'
  | 'ADMINISTRACION_Y_NEGOCIOS_DIGITALES'
  | 'ADMINISTRACION_Y_NEGOCIOS_SOSTENIBLES'
  | 'BUSINESS_ANALYTICS';